requirements.txt
black
bs4
fastparquet
flake8
ipywidgets
isort
matplotlib
numpy
pandas
pyarrow 
python-dotenv
requests
tqdm
